fix(did-web): move the fetch integration tests into src/ so wasm finds them
test:cross:integration failed on every PR in the stack: all four fetch tests
reported "Test was skipped in wasm. This usually means the test wasn't compiled
with the right cfg flags."

The cause is structural. `#[dialog_common::test]` spawns the wasm inner test as
`cargo test ... --lib <name>` — the `--lib` is hardcoded — so it only ever
searches the library target. The tests lived in `tests/fetch_server.rs`, a
separate integration-test target, which `--lib` can never reach. Every existing
cross-target test in this workspace lives in `src/` for exactly this reason;
putting them in `tests/` was the mistake.

Moved them to `src/helpers/test.rs` (a `#[cfg(test)]` module), and added
`wasm-bindgen` as a wasm-target dependency: the macro's inner-test expansion
references `wasm_bindgen`, which was only a dev-dependency here, so the inner
compile failed to resolve it. `dialog-operator`, whose cross-target tests work,
declares it the same way.

Verified by simulating the exact inner-run compile
(`--config target.wasm32...rustflags=["--cfg","dialog_test_wasm_integration"]
--target wasm32-unknown-unknown --lib --features ...`): it now builds, and all
four test names are present in the resulting wasm binary.

Native: 49 passed. Clippy --all-features clean.
